Branch riccardo

+++ shelfy/app/src/main/java/com/example/shelfy/data/remote/responses/RetailPrice.kt
@@ -1,6 +1,6 @@
 package com.example.shelfy.data.remote.responses
 
 data class RetailPrice(
-    val amountInMicros: Int = 0,
+    val amountInMicros: Long = 0,
     val currencyCode: String = ""
 )
